Create lib sub-folder to separate "pure" modules
